[context] Split ModelFeatures into its own header
Every ModelFeatures field is transformer-model topology - qk_norm, mlp kind,
norm placement, sliding window, PLE, lm-head scheme - and not one is a device
capability. Defining it in context.h made a model include the Context header to
declare what it is, which couples the model layer to the backend layer and
invites exactly the misreading that these are hardware flags.

ModelFeatures, MlpKind/NormStyle/LmHeadKind and their toString()s move verbatim
into a dependency-free `model_features.h` (<sstream>/<string> only); context.h
includes it so the resolver signature still compiles, and it is added to the
installed header set. Pure code motion - no field, no default and no caller
changes.

ExecPlan, GemmPath and resolveExecPlan stay beside DeviceCaps: the resolver is
the seam where the two vocabularies meet and genuinely needs both. Its doc
comment now states plan ownership - once the resolver is authoritative the
resolved plan belongs to the compiled model, one per model instance, while a
Context owns only its DeviceCaps - and no longer claims the queue-sync cell
"has no probe", since the device's fine-grain-SVM capability is exactly that
probe. Internal work-item tags and env-var names this change set does not
define are dropped from the comments it adds.
